Short delay after game ends, koth refinements
After the game ends, there is a 1.5 second delay before players get teleported back to the lobby.
Refined the bossbar in King of the Hill (looks better, less confusing, use with newest resource pack). If the game goes to overtime, it ends as soon as a team as the advantage, instead of them having to have complete control.