Started updating pack for 1.19.4
Changed pack format to 11.
I also started using data storage (nbt arrays), so far to rework the technicalities for bolt q, and a universal damage queue (only implemented bolt q so far). The new system, while  being a little slower (not significant), is more abstract and is not bound to the 12 player limit. I'll try to remove that limit with these new techniques if possible.