Improved start of game and players leaving
Players get only teleported every 10 ticks, hopefully this will fix a bug where some players would be stuck under the map during the game start phase.
Changed the firstjoin advancement and move_player function so when a player leaves and joins during a game, it wouldnt affect player in the game.
Removed the delay of the /clear when being moved to lobby, so far without issues, however this will be reverted if problems come up again (in the past it would sometimes look like players were still wearing their armor in lobby)

Co-Authored-By: sling233 <102863996+sling233@users.noreply.github.com>