Tank changes
F is now what R was (the projectile shield).
R now works as follows:
The Tank passively accumulates the damage taken, up to 10 hearts. When it's over 2 hearts, the tank can use R on a teammate to heal them by the amount of damage taken (it resets when used).
The Tank might be a bit over tuned right now, but balancing comes later.