Overhauled Map Management and Map Settings (ft. map editing)
Added a weather option for maps.
Added a launchpad toggle for maps.
Maps can be edited now!
Map management is entirely done through the map selector section of the admin book.
All maps updated because of a name change due to technical reasons.
Re-added version check on reload for maps and the lobby.