Improve (and fix) version handling
Fix the pack not setting the version number when installing for the first time. This led to it thinking the version changed when reloading for second time.
removeobj now removes version objective as well
Technical: checking if the pack is installed of the fist time is done via "Global version" score instead of "Global click"