fix disable reload message, improve pack updating experience
Map converting converts name for Official maps (so its not {"text":"Splat"}). Not sure if its possible for non official maps.
Pack update prompt comes before lobby set prompt. When updating from 6.0.0, the lobby data is not found (so the prompt gets sent), which is fixed by updating the pack, so that order makes more sense.
When updating from before v607, players in the lobby get cleared since most of the items changed.